requirements.txt
fastapi
uvicorn[standard]
jinja2
pandas
python-dotenv